+++ CHANGES.md
@@ -9,6 +9,7 @@ Changes in 3.9-beta
 - Move Help_*.php files from root to their corresponding locale/[code].utf8/ folder
 - Check Moodle URL and token are valid in plugins/Moodle/config.inc.php, thanks to @abogadeer
 - Fix #241 assignments of other teachers appear in StudentAssignments.fnc.php
+- SaveTemplate() $staff_id param: use 0 for default template in Template.fnc.php
 
 Changes in 3.8
 --------------